rename digestcontext.hpp into digest_context.hpp
